# api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API for SHOP Website",
    "description": "This is a REST API application made by ZDiDane for SHOP Website. \n\nIf you want to use this API, please get csrfToken first, then add csrfToken to header of request. \n\nIf you want to use API for Admin, please add bearAuth to header of request.",
    "contact": {
      "name": "ZDiDane",
      "email": "didannguyen17@gmail.com",
      "url": "https://didan.id.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    },
    {
      "url": "https://localhost:3000/api"
    },
    {
      "url": "http://shop.didan.id.vn/api"
    },
    {
      "url": "https://shop.didan.id.vn/api"
    }
  ],
...
  "paths": {
...
    "/chat": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get list chat room",
        "description": "Endpoint to get list chat room",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "csrfToken": [],
            "bearAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Chat"
        ],
        "description": "Endpoint to create chat message",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "csrfToken": [],
            "bearAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "description": "Id of product, can be null",
                    "type": "string",
                    "required": false
                  },
                  "url": {
                    "description": "Url of product, can be null",
                    "type": "string",
                    "required": false
                  },
                  "message": {
                    "description": "Message of chat",
                    "type": "string",
                    "required": true
                  },
                  "receiver": {
                    "description": "Id of receiver",
                    "type": "string",
                    "required": true
                  }
                },
                "required": [
                  "message",
                  "receiver"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}